Fix bugs in Authority

+++ src/main/java/com/example/balancing/models/user/Authority.java
@@ -8,15 +8,24 @@ import lombok.Data;
 @Data
 public class Authority {
 
+    /**
+     * Уникальный идентификатор роли.
+     */
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
 
+    /**
+     * Связь с идентификатором пользователем.
+     */
     @OneToOne(fetch = FetchType.LAZY, cascade = CascadeType.REMOVE)
     @JoinColumn(name = "user_id")
     private User user;
 
+    /**
+     * Строковое представление роли: ROLE_MODERATOR, ROLE_USER, DELETED_ACCOUNT ...
+     */
     @Column(name = "authority")
     private String authority;
